ShortURL: Move validateRelativePath to its own validate.go file
Pairs the validation logic with the existing validate_test.go
and keeps app.go focused on app setup and routing.
